Updated CSS and HTML
As a class we went over the code that we did at first with Jedi, but instead Clawson was teaching it. I made some changes to the form box such as the comment box which was not positioned properly as well as the spinner which was also not positioned properly. Some rearranging was done to the form items that weren't aligned correctly.