Final Commit. Ready for Review
Finished the exercise for this unit. We added a submit button and a cancel button which would allow the user to submit their form or to reset it by hitting the cancel button. We also added Regular Expression for the Zip Code and the Phone Number which only allows a certain way to be submitted.